refactor: Consolidate applyDatasetChartDefaults tests into hydrateExplore.test.ts
Move all 15 test cases for applyDatasetChartDefaults from separate test file
into the main hydrateExplore.test.ts file where the function is defined.

This follows standard practice of keeping tests close to their source code
for better maintainability and discoverability.
